requirements.txt
chromadb
pytesseract
streamlit
pymupdf
sentence-transformers
google-generativeai